bench that recreates prod issue
